// src/services/model/deliveryHistoryEntityPropertyChange.ts
/**
 * Generated by orval v6.29.1 🍺
 * Do not edit manually.
 * WarehouseManagement.Api
 * OpenAPI spec version: 1.0
 */

export type DeliveryHistoryEntityPropertyChange =
  (typeof DeliveryHistoryEntityPropertyChange)[keyof typeof DeliveryHistoryEntityPropertyChange]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DeliveryHistoryEntityPropertyChange = {
  StartedProcessing: 'StartedProcessing',
  FinishedProcessing: 'FinishedProcessing',
  Status: 'Status',
  ZoneId: 'ZoneId',
  AdminComment: 'AdminComment'
} as const